% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{ag_fetch}
\alias{ag_fetch}
\title{Fetch Data Used in Autograder}
\usage{
ag_fetch(problem = NULL, when = c("last-day", "last-week", "last-month",
  "start-day"), from = Sys.Date(), to = Sys.date(), save_dir = getwd())
}
\arguments{
\item{problem}{A \code{string} indicating the problem to obtain
data for.}

\item{when}{A \code{string} containing either:
\code{last-day}, \code{last-week}, \code{last-month}, or
\code{'start-day'}. If present, then parameters \code{from} and
\code{to} are ignored.}

\item{from}{A \code{date} in \code{yyyy-mm-dd} format that represents the
starting time. This parameter is ignored if \code{when} is given.}

\item{to}{A \code{date} in \code{yyyy-mm-dd} format that represents the
end time. This parameter is ignored if \code{when} is given.}

\item{save_dir}{A \code{string} indicating the save path to use.}
}
\value{
A \code{data.frame} containing problem specific data.
}
\description{
Downloads and loads into R the problem specific data.
}
\examples{
\dontrun{
## Download all the data in one file
ag_fetch(problem = "twitter-proj", when = "start-day")

## Download data from the last day
ag_fetch(problem = "twitter-proj", when = "last-day")

## Download data for specific time intervals
ag_fetch(problem = "twitter-proj",
              from = "2016-10-09", to = "2016-10-10")
}
}

