% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autograder.R
\docType{package}
\name{autograde-package}
\alias{autograde}
\alias{autograde-package}
\title{Tools for use with Autograder}
\description{
Provides local and remote computational components for use with
the autograder platform.
}

