% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistician-names.R
\name{create_stat_names}
\alias{create_stat_names}
\title{Famous Statistician Pseudonyms}
\usage{
create_stat_names(n, replace = FALSE)
}
\arguments{
\item{n}{An \code{integer} providing the number of last names to draw.}

\item{replace}{A \code{boolean} indicating if the last name can be used again.}
}
\value{
A \code{character vector} containing the statistician last name appended
by a random number.
}
\description{
Generate a list of statistician last names to be used to mask student
identities.
}
\examples{
create_stat_names(5)

create_stat_names(10, replace = TRUE)
}

