% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_agfile.R
\name{gen_ag_csv}
\alias{gen_ag_csv}
\title{Autograde Single Column CSV File Format}
\usage{
gen_ag_csv(yhat, file.name = "ag_sub")
}
\arguments{
\item{yhat}{A \code{vector} of predicted \eqn{y} values.}

\item{file.name}{A \code{string} to name the file (without a file extension).}
}
\value{
A \code{.csv} with observations on each new line written to the working
directory given by \code{getwd()}.
}
\description{
Creates a valid autograde file when supplied with a single prediction vector
}
\examples{

# Set seed for reproducibility
set.seed(1337)

# Simulate values
n = 100
beta = 2.5
x = beta*rnorm(n, 0, 1)
y = x + rnorm(n, 1, 2)

# Create a regression model
mod = lm(y~x)

## Predict under model
yhat = predict(mod)

gen_ag_csv(yhat,file.name="ag_org")

## Predict under new data
new = data.frame(x = seq(-3, 3, 0.5))
yhat2 = predict(mod, new)

gen_ag_csv(yhat2, file.name="ag_new")
}

