% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_ag_problems.R
\name{make_ci_pred}
\alias{make_ci_pred}
\title{Make CI Predictions}
\usage{
make_ci_pred(ci90, ci95, ci99)
}
\arguments{
\item{ci90, ci95, ci99}{Two \code{numeric} values for each of the parameters.}
}
\value{
A \code{.csv} with observations on each new line written to the working
directory given by \code{getwd()}.
}
\description{
Creates a valid autograde file for CI Mu Problem
}
\examples{
\dontrun{
# Make CI AG File
make_ci_pred(ci90 = c(-1,1), ci95 = c(-1.5,1.5), ci99 = c(-2.5,2.5))
}
}

