% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_ag_problems.R
\name{make_hourly_pred}
\alias{make_hourly_pred}
\title{Make Hourly Predictions}
\usage{
make_hourly_pred(hourly)
}
\arguments{
\item{hourly}{A \code{numeric} vector with 24 values.}
}
\value{
A \code{.csv} with observations on each new line written to the working
directory given by \code{getwd()}.
}
\description{
Creates a valid autograde file for hourly  project
}
\examples{
\dontrun{
# Make some predictions
hourly = 1:24
make_hourly_pred(hourly)
}
}

