% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_ag_problems.R
\name{make_pred}
\alias{make_pred}
\title{Make Predictions}
\usage{
make_pred(x)
}
\arguments{
\item{x}{Ten \code{numeric} values.}
}
\value{
A \code{.csv} with observations on each new line written to the working
directory given by \code{getwd()}.
}
\description{
Creates a valid autograde file for the prediction problems
}
\examples{
\dontrun{
# Make some predictions
make_pred(x = 1:10)
}
}

