% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_ag_problems.R
\name{make_weather_pred}
\alias{make_weather_pred}
\title{Make Weather Predictions}
\usage{
make_weather_pred(hi, lo, mean)
}
\arguments{
\item{hi, lo, mean}{A single \code{numeric} value for each of the parameters.}
}
\value{
A \code{.csv} with observations on each new line written to the working
directory given by \code{getwd()}.
}
\description{
Creates a valid autograde file for Weather Project
}
\examples{
\dontrun{
# Make some predictions
make_weather_pred(hi = 1, lo = 2, mean = 3)
}
}

