% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specific_ag_problems.R
\name{make_web_pred}
\alias{make_web_pred}
\title{Make Web Predictions}
\usage{
make_web_pred(Rcpp, ggplot2, stringr, dplyr)
}
\arguments{
\item{Rcpp, ggplot2, stringr, dplyr}{A single \code{numeric} value for each of the parameters.}
}
\value{
A \code{.csv} with observations on each new line written to the working
directory given by \code{getwd()}.
}
\description{
Creates a valid autograde file for web analytics project
}
\examples{
\dontrun{
# Make some predictions
Rcpp = 1
ggplot2 = 2
stringr = 3
dplyr = 4

make_web_pred(Rcpp, ggplot2, stringr, dplyr)
}
}

